+++ frontend/src/components/components/RequirementsWarning.jsx
@@ -1,0 +1,57 @@
+import React from 'react';
+
+const RequirementsWarning = ({ 
+  canGenerate, 
+  isLoading, 
+  connectionStatus, 
+  measure, 
+  filtersCount 
+}) => {
+  if (canGenerate || isLoading) return null;
+
+  const getMissingRequirements = () => {
+    const missing = [];
+    
+    if (connectionStatus === 'failed') {
+      missing.push('Backend connection required');
+    }
+    
+    if (!measure) {
+      missing.push('Select a Measure (Players or Teams)');
+    }
+    
+    if (filtersCount === 0) {
+      missing.push('Add at least one Filter');
+    }
+    
+    return missing;
+  };
+
+  const missingRequirements = getMissingRequirements();
+
+  if (missingRequirements.length === 0) return null;
+
+  return (
+    <div className="bg-yellow-50 border border-yellow-200 rounded-lg p-4">
+      <div className="flex">
+        <div className="flex-shrink-0">
+          <svg className="h-5 w-5 text-yellow-400" viewBox="0 0 20 20" fill="currentColor">
+            <path fillRule="evenodd" d="M8.257 3.099c.765-1.36 2.722-1.36 3.486 0l5.58 9.92c.75 1.334-.213 2.98-1.742 2.98H4.42c-1.53 0-2.493-1.646-1.743-2.98l5.58-9.92zM11 13a1 1 0 11-2 0 1 1 0 012 0zm-1-8a1 1 0 00-1 1v3a1 1 0 002 0V6a1 1 0 00-1-1z" clipRule="evenodd" />
+          </svg>
+        </div>
+        <div className="ml-3">
+          <h3 className="text-sm font-medium text-yellow-800">Requirements Missing</h3>
+          <div className="mt-2 text-sm text-yellow-700">
+            <ul className="list-disc list-inside space-y-1">
+              {missingRequirements.map((requirement, index) => (
+                <li key={index}>{requirement}</li>
+              ))}
+            </ul>
+          </div>
+        </div>
+      </div>
+    </div>
+  );
+};
+
+export default RequirementsWarning;
